refactor(app): split persistence helpers
