+++ packages/app/src/utils/persist-local-storage.ts
@@ -1,0 +1,228 @@
+import type { SyncStorage } from "@solid-primitives/storage"
+
+const LOCAL_PREFIX = "pawwork."
+const LEGACY_LOCAL_PREFIX = "opencode."
+// The legacy prefix intentionally covers old workspace, global, and settings keys.
+const EVICT_PREFIXES = [LOCAL_PREFIX, LEGACY_LOCAL_PREFIX]
+const fallback = new Map<string, boolean>()
+
+const CACHE_MAX_ENTRIES = 500
+const CACHE_MAX_BYTES = 8 * 1024 * 1024
+
+type CacheEntry = { value: string; bytes: number }
+const cache = new Map<string, CacheEntry>()
+const cacheTotal = { bytes: 0 }
+
+function cacheDelete(key: string) {
+  const entry = cache.get(key)
+  if (!entry) return
+  cacheTotal.bytes -= entry.bytes
+  cache.delete(key)
+}
+
+function cachePrune() {
+  for (;;) {
+    if (cache.size <= CACHE_MAX_ENTRIES && cacheTotal.bytes <= CACHE_MAX_BYTES) return
+    const oldest = cache.keys().next().value as string | undefined
+    if (!oldest) return
+    cacheDelete(oldest)
+  }
+}
+
+function cacheSet(key: string, value: string) {
+  const bytes = value.length * 2
+  if (bytes > CACHE_MAX_BYTES) {
+    cacheDelete(key)
+    return
+  }
+
+  const entry = cache.get(key)
+  if (entry) cacheTotal.bytes -= entry.bytes
+  cache.delete(key)
+  cache.set(key, { value, bytes })
+  cacheTotal.bytes += bytes
+  cachePrune()
+}
+
+function cacheGet(key: string) {
+  const entry = cache.get(key)
+  if (!entry) return
+  cache.delete(key)
+  cache.set(key, entry)
+  return entry.value
+}
+
+function fallbackDisabled(scope: string) {
+  return fallback.get(scope) === true
+}
+
+function fallbackSet(scope: string) {
+  fallback.set(scope, true)
+}
+
+function quota(error: unknown) {
+  if (error instanceof DOMException) {
+    if (error.name === "QuotaExceededError") return true
+    if (error.name === "NS_ERROR_DOM_QUOTA_REACHED") return true
+    if (error.name === "QUOTA_EXCEEDED_ERR") return true
+    if (error.code === 22 || error.code === 1014) return true
+    return false
+  }
+
+  if (!error || typeof error !== "object") return false
+  const name = (error as { name?: string }).name
+  if (name === "QuotaExceededError" || name === "NS_ERROR_DOM_QUOTA_REACHED") return true
+  if (name && /quota/i.test(name)) return true
+
+  const code = (error as { code?: number }).code
+  if (code === 22 || code === 1014) return true
+
+  const message = (error as { message?: string }).message
+  if (typeof message !== "string") return false
+  if (/quota/i.test(message)) return true
+  return false
+}
+
+type Evict = { key: string; size: number }
+
+function evict(storage: Storage, keep: string, value: string) {
+  const total = storage.length
+  const indexes = Array.from({ length: total }, (_, index) => index)
+  const items: Evict[] = []
+
+  for (const index of indexes) {
+    const name = storage.key(index)
+    if (!name) continue
+    if (!EVICT_PREFIXES.some((prefix) => name.startsWith(prefix))) continue
+    if (name === keep) continue
+    const stored = storage.getItem(name)
+    items.push({ key: name, size: stored?.length ?? 0 })
+  }
+
+  items.sort((a, b) => b.size - a.size)
+
+  for (const item of items) {
+    storage.removeItem(item.key)
+    cacheDelete(item.key)
+
+    try {
+      storage.setItem(keep, value)
+      cacheSet(keep, value)
+      return true
+    } catch (error) {
+      if (!quota(error)) throw error
+    }
+  }
+
+  return false
+}
+
+function write(storage: Storage, key: string, value: string) {
+  try {
+    storage.setItem(key, value)
+    cacheSet(key, value)
+    return true
+  } catch (error) {
+    if (!quota(error)) throw error
+  }
+
+  try {
+    storage.removeItem(key)
+    cacheDelete(key)
+    storage.setItem(key, value)
+    cacheSet(key, value)
+    return true
+  } catch (error) {
+    if (!quota(error)) throw error
+  }
+
+  const ok = evict(storage, key, value)
+  return ok
+}
+
+export function localStorageWithPrefix(prefix: string): SyncStorage {
+  const base = `${prefix}:`
+  const scope = `prefix:${prefix}`
+  const item = (key: string) => base + key
+  return {
+    getItem: (key) => {
+      const name = item(key)
+      const cached = cacheGet(name)
+      if (fallbackDisabled(scope)) return cached ?? null
+
+      const stored = (() => {
+        try {
+          return localStorage.getItem(name)
+        } catch {
+          fallbackSet(scope)
+          return null
+        }
+      })()
+      if (stored === null) return cached ?? null
+      cacheSet(name, stored)
+      return stored
+    },
+    setItem: (key, value) => {
+      const name = item(key)
+      if (fallbackDisabled(scope)) return
+      try {
+        if (write(localStorage, name, value)) return
+      } catch {
+        fallbackSet(scope)
+        return
+      }
+      fallbackSet(scope)
+    },
+    removeItem: (key) => {
+      const name = item(key)
+      cacheDelete(name)
+      if (fallbackDisabled(scope)) return
+      try {
+        localStorage.removeItem(name)
+      } catch {
+        fallbackSet(scope)
+      }
+    },
+  }
+}
+
+export function localStorageDirect(): SyncStorage {
+  const scope = "direct"
+  return {
+    getItem: (key) => {
+      const cached = cacheGet(key)
+      if (fallbackDisabled(scope)) return cached ?? null
+
+      const stored = (() => {
+        try {
+          return localStorage.getItem(key)
+        } catch {
+          fallbackSet(scope)
+          return null
+        }
+      })()
+      if (stored === null) return cached ?? null
+      cacheSet(key, stored)
+      return stored
+    },
+    setItem: (key, value) => {
+      if (fallbackDisabled(scope)) return
+      try {
+        if (write(localStorage, key, value)) return
+      } catch {
+        fallbackSet(scope)
+        return
+      }
+      fallbackSet(scope)
+    },
+    removeItem: (key) => {
+      cacheDelete(key)
+      if (fallbackDisabled(scope)) return
+      try {
+        localStorage.removeItem(key)
+      } catch {
+        fallbackSet(scope)
+      }
+    },
+  }
+}
