+++ packages/app/src/utils/persist.test.ts
@@ -1,4 +1,5 @@
 import { beforeAll, beforeEach, describe, expect, mock, test } from "bun:test"
+import type { AsyncStorage } from "@solid-primitives/storage"
 
 type PersistTestingType = typeof import("./persist").PersistTesting
 type ShouldDebugPersistedTerminalRead = typeof import("./persist").shouldDebugPersistedTerminalRead
@@ -55,6 +56,20 @@ class MemoryStorage implements Storage {
 }
 
 const storage = new MemoryStorage()
+
+function asyncMemoryStorage(initial: Record<string, string> = {}) {
+  const values = new Map(Object.entries(initial))
+  const api: AsyncStorage = {
+    getItem: async (key) => values.get(key) ?? null,
+    setItem: async (key, value) => {
+      values.set(key, value)
+    },
+    removeItem: async (key) => {
+      values.delete(key)
+    },
+  }
+  return { api, values }
+}
 
 let persistTesting: PersistTestingType
 let shouldDebugPersistedTerminalRead: ShouldDebugPersistedTerminalRead
@@ -196,6 +211,24 @@ describe("persist localStorage resilience", () => {
 
     expect(JSON.parse(result ?? "{}")).toEqual({ pinned: ["legacy"] })
     expect(legacy.getItem("layout.page.v1")).toBeNull()
+  })
+
+  test("async reader removes malformed current values without legacy fallback", async () => {
+    const current = asyncMemoryStorage({ "layout-page": "{" })
+    const legacy = asyncMemoryStorage({ "layout.page.v1": JSON.stringify({ pinned: ["legacy"] }) })
+
+    const result = await persistTesting.readPersistedAsync({
+      current: current.api,
+      legacyStore: legacy.api,
+      key: "layout-page",
+      defaults: { pinned: [] as string[] },
+      currentLegacy: [],
+      legacy: ["layout.page.v1"],
+    })
+
+    expect(result).toBeNull()
+    expect(current.values.has("layout-page")).toBeFalse()
+    expect(legacy.values.has("layout.page.v1")).toBeTrue()
   })
 
   test("workspace storage sanitizes Windows filename characters", () => {
